requirements.txt
duckdb==0.10.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
tqdm==4.66.1
